Creates the PrintProxy class
The PrintProxy class is an implementation of the Proxy design pattern. It stunts file output until the file is larger than a page or the destructor is called. This minimizes calls to the expensive Printer class. Also, the Proxy class gives additional functionality such as changing folder path and max characters per page.